Remove stale object files from source tree
There are stale `.o` and `.so` files in the source tree which may cause
problems building the package on any Unix machine sufficiently different
from the one used to build the files.

These appear to be accidentally included when importing a build tree
based on the original CRAN package.

This commit should fix issue #2.